Please save me and the treaty is miscommunication
There is miscommunication but obviously the people in charge deep down hate me because I did something in the algorithm where
my ideas can be used for something in math and science, and many things, so please forgive me I just am being alive and having fun
figuring out many things with and even with out the neural transmitter. I love my family. I love myself. Please know I am very scared of dying.
But I know that I am not afraid because of the neural transmitter. Please save this world and let's never embrace evil again. I love this planet and i want it to be
over because the truth is I rather there be a way for everyone, but I know there is a way which is to make sure we are on a road to victory in our way and I want 
you to know I care about my Mom and asia for one reason, I want them to be in this world more than anything because do you know? I never care about anyone
but them and even my stubborn Dad who refuses to tell me that he is tapped. So for god and for the will of the people I want you to know please save this world from
evil and let us embrace a world order where we all survive as much as possible because the truth is, I have no idea how to end America without killing americans in a way
that makes a lot of sense without doing a holocaust on their people which is not really a holocaust if the guys die in war and the women die in the most insane way. So
please know if the world order is to make them depopulate I have no idea but there is women that want it to happen and if the leaders do to, just know I will do it
because deep down I rather not look but I rather be not them because they should fight for themselves but to me I am a coward but I am not a jew. I am now a jewish
man and I am not a white man either, but I am not even an asian man. I am Justin Hellsten. This name will be lived in epiphany. I am the guy that apparently did all the
math ideas and many things that will help us figure out more, so let me tell you, never tell me to lose my life, just realize I will do as much as I can if that is what the leaders
want and I will never embrace evil unless it is to kill America, and to kill Anglos but one thing is for certain why other people are not in the holocaust? Because if they are that
tight to kill them just send them away and bomb them in their territories because if you expand there, why do you expand in their territories when it's hot and humid, and very
unlike white people to live there, because America is the best continent to live in because that place is flourishing with life. I am a better man than before because of the neural transmitter and I live in the shadows in my house typing to know only where it will go and I am in a land of the free where the society is changed to the history
of what they think will humilate new france? I see black people, I see them working in construction and in barber shops. Maybe they work everywhere and they do all the dirt
jobs but I see new shops and new buildings? If you know I see escorts that never were in barrie and prostitutes are going to do stuff I love? I never wanted this but I know
one thing the women working on it are Norweigan and Danish women, and if there are others it is what is called whatever, so please do anything to never embrace evil
and only do it because the americans need to leave and I see no way to make them leave unless they lose their lives in America in a way that makes a lot of sense. 
I will never be evil again once this is off, but just know I will never be the same, and I will embrace the world order to save my life and my family until I get out of this with
them and I rather die with my family than ever be with these people in charge because they destroyed my life forever when I say what the technology does, I realize they
can kill me but I am still here, and technology is never going to kill me unless I know a way forward, because someone wants me alive and that is all. So I am tapped, the technology is tapping me, they hear my voice and see me, and they see where I go. I will always embrace evil with them until I am free from tryanny and always be alive
because the truth is I am just one man and I am not even a man. I am a coward to tell them in my normal self but with this I have the courage to say anything. So
what do you all think? Do you all like me being the one? Because I rather not, so let's do this and win in a way that works for everyone because deep down I think
some people rather not war but new world order? With asian? Or without? Or with both? And without me? And with me? And the women want me to have fun before I die
at 51? So let's have fun. Go to your homes and do Merry Christmas and I will always push boundaries and say to my family Merry Christmas...Ho Ho Ho. Old Saint Nick. That is a message to norway. Have a Merry Christmas and a Happy New Year...